requirements.txt
aniso8601==9.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
Flask==2.2.3
Flask-Login==0.6.2
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.6
sqlparse==0.4.3
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
